# bff-openapi.yaml
openapi: 3.0.3
info:
  title: my-token BFF API
  version: "1.0.0"
  description: >
    The BFF's session-authenticated JSON surface (milestone-3
    `.chief/milestone-3/_contract/API.md`), consumed only by the SPA
    running in the owner's own browser. Every request authenticates over
    a signed, HttpOnly session cookie, resolved by
    `internal/transport/bff`'s own session middleware
    (`bff.RequireJSONSession`) — not by anything in this spec: no
    `security` scheme is declared here on purpose, for the exact same
    reason `openapi.yaml`'s own header comment gives for that file: this
    spec's job is request/response *shape* validation (`gin-middleware`),
    not auth. A second, independent file from `openapi.yaml` on purpose
    (`_contract/API.md` "Two specs, not one") — the public API
    (Bearer-authenticated, agent-facing) and this BFF surface
    (session-authenticated, owner-facing browser only) are two disjoint
    proof-of-identity mechanisms by construction (I2/I12), and a single
    shared spec file would blur that boundary even though the code
    enforces it.

    Deliberately absent from this surface: `POST /todos` aside, there is
    no key issuance or rotation endpoint here (no `POST /keys`, no
    rotate) — same reasoning `openapi.yaml`'s own `/keys/{id}` comment
    gives: a raw key can't be returned safely over an HTTP response, so
    issuance and rotation stay CLI-only regardless of which surface is
...

...
servers:
  - url: /api/bff
paths:
...
  /todos:
    get:
      operationId: listTodos
      summary: List every todo.
      description: >
        milestone-4: todos are a shared collection (`_rules/_contract/
        DATA_MODEL.md`/`GOAL.md`'s Ownership model decision) — I3 no
        longer applies to this domain, on either surface. Returns every
        todo, created_at descending, unpaginated — same shape as
        `_rules/_contract/API.md`'s `GET /api/v1/todos`, not just the
        session owner's own.
      responses:
        "200":
          description: Every todo.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TodoList"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createTodo
      summary: Create a todo, attributed to the caller.
      description: >
        `createdBy` is always the resolved session owner — never accepted
        from the body (I1, same rule as the public API, different auth
        mechanism) — but is attribution only, never access-scoping
        (GOAL.md). `status` always starts `open`. The `created` event
        this produces is never client-specifiable as a distinct write
        (I16) — it only ever happens as this endpoint's own side effect.
        `clientRequestId` makes this idempotent the same way
        `POST .../events` is (I19). Calls
        `internal/domain/todo.Service.CreateTodo` directly — the same
        service instance and method the public API's handler calls
        (`_rules/_standard/ARCHITECTURE.md`'s shared-service-layer rule).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTodoRequest"
      responses:
        "201":
          description: The created todo.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Todo"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /keys:
    get:
      operationId: listKeys
      summary: List every agent's non-revoked API keys (I21).
      description: >
        Every `role='agent'` user's non-revoked keys (`revoked_at IS
        NULL`), not just the session owner's own — the session owner
        never holds a key at all (`role='owner'` can't be issued one,
        I2), so scoping this to the caller would always return empty,
        which was milestone-2/3's actual bug. Regardless of expiry: an
        expired-but-unrevoked key still shows up so the owner can see it
        needs rotating. Each row carries the owning agent's `handle`
        (never null — only rows a real users JOIN produced ever reach
        this response), so the owner can tell whose key is whose before
        revoking one.
      responses:
        "200":
          description: Every agent's non-revoked keys.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiKeyList"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /keys/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    delete:
      operationId: revokeKey
      summary: Revoke any agent's API key by id (I21).
      description: >
        Sets `revoked_at`. Not scoped to a "caller's own" key — there is
        no such thing on this owner-only surface (I2) — any agent's key
        id may be revoked, session-gated only. An unknown id is
        `404 not_found`, the same "absence, not permission" shape every
        other owner-scoped-by-nothing-left-to-scope lookup on this
        surface uses. No `POST /keys` and no rotate endpoint anywhere on
        this surface, deliberately — a raw key can't be returned safely
        over an HTTP response, so issuance/rotation stay CLI-only
        (`cmd/issue-key`) regardless of which surface is asking.
      responses:
        "204":
          description: Revoked.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /usage/summary:
    get:
      operationId: getUsageSummary
      summary: Usage totals and one-dimension breakdown for a time window.
      description: >
        story-1/ticket-14: the console's own read surface over
        `usage_events` (`_rules/../story-1/_contract/contract.md`'s
        "API surface" section — wire field names below are copied
        verbatim from that contract, including its `snake_case`
        `group_by`/`reporting_installs`, which deliberately breaks from
        this surface's own camelCase convention elsewhere; not
        "corrected" here since the contract is this ticket's literal
        input, not something to reopen). `window` selects the
        [start, now) range (`today`/`week`/`month` are calendar-anchored
        at UTC midnight/Monday/1st; `5h`/`24h` are rolling); `group_by`
        picks which dimension the `breakdown` array is keyed by.
        `totals.tokens` sums all four raw token counters per event
        (input + output + cache_read + cache_creation) — a single
        display number, distinct from ticket 9's "never summed into
        each other" rule about the *stored* per-field columns, which
        this endpoint only ever reads, never rewrites.
        `reporting_installs` is the count of distinct `machine` values
        with at least one event inside the selected window (not
        lifetime) — the goal's partial-view caveat, scoped to what the
        rest of this response actually covers.


        `lifetime` is a sixth value, additive to the contract's own
        fixed five (`_contract/contract.md`'s API surface section names
        exactly `5h|24h|today|week|month`) — flagged as a build-time
        decision in ticket 14's own report, not silently added: the
        ticket's own mockup spec names a "lifetime cost" summary tile,
        which none of the five contract windows can answer (each has a
        bounded start). Query-string-additive only — `GET
        /usage/windows`' own fixed table is untouched, still exactly
        those five, in that order. `lifetime` covers every event ever
        ingested up to now, with no left bound.
      parameters:
        - name: window
          in: query
          required: true
          schema:
            type: string
            enum: [5h, 24h, today, week, month, lifetime]
        - name: group_by
          in: query
          required: true
          schema:
            type: string
            enum: [actor, path, machine]
      responses:
        "200":
          description: Totals, a breakdown by the requested dimension, and the window-scoped reporting-install count.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsageSummary"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /usage/windows:
    get:
      operationId: getUsageWindows
      summary: Turns/tokens/cost for each of the fixed 5h/24h/today/week/month windows.
      description: >
        story-1/ticket-14: backs the console's time-window table — one
        row per fixed window, each computed the same way
        `getUsageSummary`'s own `window` parameter computes its range,
        with no `group_by` (each row is this service's whole reporting
        surface's totals for that window, not split by dimension).
      responses:
        "200":
          description: One row per fixed window, in 5h, 24h, today, week, month order.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsageWindows"
        "401":
          $ref: "#/components/responses/Unauthorized"
components:
  schemas:
...
    Todo:
      type: object
      required:
        [
          id,
          title,
          status,
          assigneeId,
          assigneeHandle,
          priority,
          dueDate,
          createdBy,
          createdAt,
          updatedAt,
        ]
      properties:
        id:
          type: string
        title:
          type: string
        status:
          type: string
          description: >
            milestone-4: replaces `done`. Fixed enum, not an
            owner-editable table (GOAL.md).
          enum: [open, in_progress, done, closed]
        assigneeId:
          type: string
          nullable: true
          description: References a user (any role).
        assigneeHandle:
          type: string
          nullable: true
          description: >
            milestone-4 fix-round (handle-exposure): assigneeId's owning
            user's handle, for display — null exactly when assigneeId is
            null. my-task's own task list/detail views show a plain handle
            (`~/gits/my-task/src/server/modules/task/task.queries.ts`'s
            `assigneeHandle` LEFT JOIN, `t.assignee`/`detail.assignee` on
            the wire), never a bare id, and task-7's report found this
            repo's own Todo had no equivalent. Additive: assigneeId is
            unchanged, still what a caller writes back
            (CreateTodoRequest.assigneeId, CreateTodoEventRequest's `to`
            for an `assigned` event) — this is display-only, openapi.yaml's
            own Todo carries the identical field for the same reason
            (my-task's agent-facing REST shows it too, not just the owner
            UI).
        priority:
          type: string
          nullable: true
          enum: [low, medium, high, urgent]
        dueDate:
          type: string
          format: date-time
          nullable: true
        createdBy:
          type: string
          description: >
            milestone-4: replaces `ownerId` — attribution only, never
            access-scoping (GOAL.md).
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    TodoList:
      type: object
      required: [todos]
      properties:
        todos:
          type: array
          items:
            $ref: "#/components/schemas/Todo"
...
    ApiKey:
      type: object
      required: [id, prefix, handle, createdAt, expiresAt]
      properties:
        id:
          type: string
        prefix:
          type: string
        handle:
          type: string
          description: >
            milestone-4 fix-round (handle-exposure): the owning agent's
            handle — my-task's own api-key-settings.tsx shows `{k.handle}`
            on every row and its revoke dialog names the agent by handle
            ("Revoke {handle}'s key?"); this endpoint (GET /api/bff/keys)
            already JOINs users for every row it returns (I21 — every key
            here belongs to a role='agent' user), so this is always
            present, never null.
        createdAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
    ApiKeyList:
      type: object
      required: [keys]
      properties:
        keys:
          type: array
          items:
            $ref: "#/components/schemas/ApiKey"
...
    CreateTodoRequest:
      type: object
      additionalProperties: false
      description: >
        `additionalProperties: false` so a stray `done` (milestone-1/2/3's
        removed field) is a `validation_error`, not silently dropped
        (`_contract/API.md`). No `status` field, deliberately (mirrors
        the public API's own `CreateTodoRequest`) — every created todo
        starts `open` regardless of what a caller asks for.
      required: [title, clientRequestId]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        assigneeId:
          type: string
          nullable: true
        priority:
          type: string
          nullable: true
          enum: [low, medium, high, urgent]
        dueDate:
          type: string
          format: date-time
          nullable: true
        clientRequestId:
          type: string
          minLength: 1
          description: >
            I19's idempotency key, on the `created` event this endpoint's
            own side effect writes — required, not optional, since
            `todo_events.client_request_id` is `NOT NULL UNIQUE` at the
            schema level for every row including `created` ones.
...
    UsageTotals:
      type: object
      additionalProperties: false
      required: [tokens, cost, turns]
      properties:
        tokens:
          type: integer
          format: int64
        cost:
          type: number
          format: double
        turns:
          type: integer
          format: int64
    UsageBreakdownRow:
      type: object
      additionalProperties: false
      required: [key, tokens, cost, turns]
      properties:
        key:
          type: string
          description: >
            The raw dimension value (an actor name, a canonicalized path,
            or a machine's `install_id`) — display shortening (contract's
            "Console display rules", `path` basename/collision handling)
            is the SPA's own job, not this endpoint's; this is always the
            ground-truth value.
        tokens:
          type: integer
          format: int64
        cost:
          type: number
          format: double
        turns:
          type: integer
          format: int64
    UsageSummary:
      type: object
      additionalProperties: false
      required: [totals, breakdown, reporting_installs]
      properties:
        totals:
          $ref: "#/components/schemas/UsageTotals"
        breakdown:
          type: array
          items:
            $ref: "#/components/schemas/UsageBreakdownRow"
          description: Every distinct key for the requested group_by within the window, cost descending.
        reporting_installs:
          type: integer
          format: int64
          description: Distinct `machine`s with at least one event inside the selected window — not lifetime.
    UsageWindowRow:
      type: object
      additionalProperties: false
      required: [window, turns, tokens, cost]
      properties:
        window:
          type: string
          enum: [5h, 24h, today, week, month]
        turns:
          type: integer
          format: int64
        tokens:
          type: integer
          format: int64
        cost:
          type: number
          format: double
    UsageWindows:
      type: object
      additionalProperties: false
      required: [windows]
      properties:
        windows:
          type: array
          items:
            $ref: "#/components/schemas/UsageWindowRow"
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            hint:
              type: string
  responses:
    BadRequest:
      description: The request violates a request-shape or validation rule.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: >
        Missing, expired, or wrong-role session cookie — same shape
        regardless of which check failed (I5). A behavior change from
        this package's own HTML `GET /` view: this JSON surface always
        answers 401, never a redirect (a `fetch` call can't follow a
        redirect the way a browser navigation does — the SPA's own
        AuthGate-equivalent hook is what turns a 401 into a redirect,
        client-side).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: >
        Unknown id, or an id that exists but isn't the caller's (I3 —
        absence, not permission).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"